requirements.txt
selenium==4.20.0
pyttsx3==2.90
speechrecognition==3.10.0
pyaudio==0.2.14
pygame==2.5.2
openai==1.12.0
spotdl==4.2.4
opencv-python==4.9.0.80
newsapi-python==0.2.7
requests==2.31.0
appopener==1.7
validate_email==1.3
customtkinter==5.2.2
keyboard==0.13.5
CTkMessagebox==2.5
validate_email==1.3
pillow=9.4.0